Consolidating code and simplifying.
