+++ distribution/policies/default.json
@@ -1,83 +1,1 @@
-{
-  "name": "default",
-    "config": {
-    "splitting": {
-      "enabled": true,
-      "threshold": 384,
-      "method": "width"
-    }
-  },
-  "ignored": [],
-  "identifiers": {
-    "dictionaries": [],
-    "person": {
-      "phEyeConfiguration": {
-        "endpoint": "http://philter-ph-eye-1:5000"
-      },
-      "personFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "age": {
-      "ageFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "creditCard": {
-      "creditCardFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "date": {
-      "dateFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "emailAddress": {
-      "emailAddressFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "ipAddress": {
-      "ipAddressFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "phoneNumber": {
-      "phoneNumberFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "ssn": {
-      "ssnFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "url": {
-      "urlFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "vin": {
-      "vinFilterStrategies": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    },
-    "zipCode": {
-      "zipCodeFilterStrategy": [{
-        "strategy": "REDACT",
-        "redactionFormat": "{{{REDACTED-%t}}}"
-      }]
-    }
-  }
-}
+{"name":"default","config":{"splitting":{"enabled":true,"threshold":384,"method":"width"},"pdf":{"redactionColor":"black","showReplacement":false,"replacementFont":"helvetica","replacementMaxFontSize":12.0,"scale":0.25,"dpi":150,"compressionQuality":1.0,"preserveUnredactedPages":false},"postFilters":{"removeTrailingPeriods":true,"removeTrailingSpaces":true,"removeTrailingNewLines":true},"analysis":{"identification":true}},"identifiers":{"all":false,"person":{"removePunctuation":false,"thresholds":{},"phEyeConfiguration":{"endpoint":"http://philter-ph-eye-1:5000","timeout":600,"maxIdleConnections":30,"labels":["Person"]},"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"age":{"ageFilterStrategies":[{"filterType":"AGE","id":"0405cc0f-bd12-409c-8dd4-1f06fc6dd2e3","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"creditCard":{"onlyValidCreditCardNumbers":true,"creditCardFilterStrategies":[{"filterType":"CREDIT_CARD","id":"39cb9e95-3538-4fde-8df9-f71deedf0c1d","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"ignoreWhenInUnixTimestamp":false,"onlyWordBoundaries":true,"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"date":{"onlyValidDates":false,"dateFilterStrategies":[{"filterType":"DATE","shiftRandom":false,"shiftDays":0,"shiftMonths":0,"shiftYears":0,"futureDates":false,"id":"0db54638-797c-4806-82e6-e4fcd6e14633","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"emailAddress":{"onlyStrictMatches":true,"onlyValidTLDs":false,"emailAddressFilterStrategies":[{"filterType":"EMAIL_ADDRESS","id":"ae944bcd-1e65-4e5c-a1f5-1d0814a542fa","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"ipAddress":{"ipAddressFilterStrategies":[{"filterType":"IP_ADDRESS","id":"cb2006d4-d7ce-4541-88fb-e2538e9fd88f","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"phoneNumber":{"phoneNumberFilterStrategies":[{"filterType":"PHONE_NUMBER","id":"fdcf6dd8-715c-4bec-b871-19deead9f5c2","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"ssn":{"ssnFilterStrategies":[{"filterType":"SSN","id":"3df4611d-fb21-47d8-a1e6-5a5502cbcd0b","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"url":{"urlFilterStrategies":[{"filterType":"URL","id":"892bc5fe-1b69-4fa9-aa6a-dc4760d97e62","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"requireHttpWwwPrefix":true,"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"vin":{"vinFilterStrategies":[{"filterType":"VIN","id":"6e0ba430-0841-4f02-af58-7d8e6bfaf730","strategy":"REDACT","redactionFormat":"{{{REDACTED-%t}}}","replacementScope":"DOCUMENT","staticReplacement":"","maskCharacter":"*","maskLength":"SAME","truncateCharacter":"*","truncateDirection":"LEADING","condition":"","alert":false,"salt":false}],"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0},"zipCode":{"requireDelimiter":false,"validate":false,"enabled":true,"ignored":[],"ignoredFiles":[],"ignoredPatterns":[],"windowSize":0,"priority":0}},"ignored":[],"ignoredPatterns":[],"graphical":{"boundingBoxes":[]}}
