+++ docs/mkdocs.yml
@@ -96,12 +96,10 @@ nav:
     - 'API Methods':
       - 'Filtering API': 'api_and_sdks/api/filtering_api.md'
       - 'Policies API': 'api_and_sdks/api/policies_api.md'
-      - 'Alerts API': 'api_and_sdks/api/alerts_api.md'
     - 'Client SDKs': 'api_and_sdks/sdks.md'
   - 'Settings': settings.md
   - 'Monitoring and Logging': 'monitoring_and_logging.md'
   - 'Advanced Features':
-      - 'Alerts': 'other_features/alerts.md'
       - 'Consistent Anonymization': 'other_features/consistent_anonymization.md'
       - 'Span Disambiguation': 'other_features/span_disambiguation.md'
       - 'Dashboard': 'other_features/dashboard.md'
