+++ example_api_calls/get-policy.sh
@@ -2,4 +2,4 @@
 
 POLICY_NAME="default"
 
-curl -X GET "http://localhost:8080/api/policies/${POLICY_NAME}" -H "Accept: application/json" | jq
+curl -s -X GET "http://localhost:8080/api/policies/${POLICY_NAME}" -H "Accept: application/json" | jq
