+++ example_api_calls/save-policy.sh
@@ -1,5 +1,5 @@
 #!/bin/bash -e
 
-curl -s -X POST "http://localhost:8080/api/policies" \
+curl -v -s -X POST "http://localhost:8080/api/policies" \
   -H "Content-Type: application/json" \
   -d @../distribution/policies/default.json
