+++ philter-app/src/main/java/ai/philterd/philter/PhilterConfiguration.java
