+++ philter-app/src/main/java/ai/philterd/philter/api/exceptions/BadRequestException.java
