+++ philter-app/src/main/java/ai/philterd/philter/api/model/ResponseSpan.java
