+++ philter-app/src/main/java/ai/philterd/philter/api/security/X509Authenticator.java
