--- philter-configuration/pom.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <artifactId>philter</artifactId>
-    <groupId>ai.philterd</groupId>
-    <version>3.3.0-SNAPSHOT</version>
-  </parent>
-  <artifactId>philter-configuration</artifactId>
-  <name>philter-configuration</name>
-  <dependencies>
-    <dependency>
-      <groupId>ai.philterd</groupId>
-      <artifactId>phileas-model</artifactId>
-      <version>${phileas.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.commons</groupId>
-      <artifactId>commons-lang3</artifactId>
-      <version>${commons.lang3.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <scope>test</scope>
-      <version>${junit.version}</version>
-    </dependency>
-  </dependencies>
-</project>
